fixup! Pause apps feature
Prevent non-system packages from spoofing the calling package
arbitrarily, as was already prevented prior to the change this is
fixing. This helps avoid unintended consequences caused by, for
example, setting the calling package to the platform package and
attempting to add or to remove system-level suspensions, which
should not be allowed.

If the caller is from a different user than the one whose suspensions
it wants to change, ensure that it has permission to interact across
users. If it is not calling from the parent user, it must have the
full permission.

Issue: calyxos#1599
Change-Id: Ifb1bbd3cc495c6e7682908d46e41cce2fa1f5cf4